stbt.MatchResult: Define __eq__, __ne__ and __hash__
So that we can use MatchResult objects with wait_until's
`consecutive_secs` parameter to wait until a match is found *and* it
stops moving. This is useful when you're waiting for an animation to
complete, for a selection to finish moving, etc.

Two MatchResult objects are considered equal if they represent the same
template (reference image), the match status (matched or not matched) is
the same, and at the same position. It doesn't matter if the objects are
for 2 different frames of video (we ignore the frame timestamp, the
frame itself, and the level of the first-pass match which can be
different depending on the content of the frame).

We also define __hash__ for backward compatibility; otherwise you would
no longer be able to store MatchResult objects in a set or dict. Note
that this changes the definition of equality: Before you could store
several frames worth of MatchResult objects in a set or dict, now many
of those might be considered identical objects in which case only one
will be stored in the set or dict. I doubt anyone is using MatchResult
objects in this way though.

Note that `FrameObject` already defines __cmp__ and __hash__.